Refactor logic to share across date features
