+++ packages/foam-vscode/src/features/open-dated-note.ts
@@ -1,41 +1,56 @@
 import {
-  window,
   workspace,
-  Uri,
-  WorkspaceConfiguration,
   ExtensionContext,
   commands,
-  Selection,
   languages,
   CompletionItemProvider,
   CompletionItem,
   CompletionItemKind
 } from "vscode";
-import { dirname, join } from "path";
-import dateFormat from "dateformat";
-import * as fs from "fs";
+import { getDailyNoteFileName, openDatedNote } from "../dated-notes";
 import { FoamFeature } from "../types";
-import { docConfig } from "../utils";
 
-const completion: CompletionItemProvider = {
-  provideCompletionItems: (document, position, token, context) => {
+interface DateSnippet {
+  snippet: string;
+  date: Date;
+  detail: string;
+}
+const foamConfig = workspace.getConfiguration("foam");
+const snippets: (() => DateSnippet)[] = [
+  () => ({
+    detail: "Insert a link to today's daily note",
+    snippet: "/day",
+    date: new Date()
+  }),
+  () => {
     const today = new Date();
-    const tomorrow = new CompletionItem(
-      "/tomorrow",
-      CompletionItemKind.Snippet
-    );
-    tomorrow.insertText = `[[${getDailyNoteFileName(
-      workspace.getConfiguration("foam"),
-      new Date(today.getFullYear(), today.getMonth(), today.getDate() + 1)
-    )}]]`;
-    tomorrow.command = {
-      command: "foam-vscode.open-dated-note",
-      title: "Open a note for the given date",
-      arguments: [
-        new Date(today.getFullYear(), today.getMonth(), today.getDate() + 1)
-      ]
+    return {
+      detail: "Insert a link to tomorrow's daily note",
+      snippet: "/tomorrow",
+      date: new Date(today.getFullYear(), today.getMonth(), today.getDate() + 1)
     };
-    return [tomorrow];
+  }
+];
+
+const completions: CompletionItemProvider = {
+  provideCompletionItems: (document, position, token, context) => {
+    const completionItems = snippets.map(snippet => {
+      const completionItem = new CompletionItem(
+        snippet().snippet,
+        CompletionItemKind.Snippet
+      );
+      completionItem.insertText = `[[${getDailyNoteFileName(
+        foamConfig,
+        snippet().date
+      )}]]`;
+      completionItem.command = {
+        command: "foam-vscode.open-dated-note",
+        title: "Open a note for the given date",
+        arguments: [snippet().date]
+      };
+      return completionItem;
+    });
+    return completionItems;
   }
 };
 
@@ -46,102 +61,8 @@ const feature: FoamFeature = {
         openDatedNote(date)
       )
     );
-    languages.registerCompletionItemProvider(
-      "markdown",
-      completion,
-      "/",
-      "/today",
-      "/tomorrow"
-    );
+    languages.registerCompletionItemProvider("markdown", completions, "/");
   }
 };
-
-async function openDatedNote(date: Date) {
-  const foamConfiguration = workspace.getConfiguration("foam");
-  const currentDate = date ? date : new Date();
-
-  const dailyNotePath = getDailyNotePath(foamConfiguration, currentDate);
-
-  const isNew = await createDailyNoteIfNotExists(
-    foamConfiguration,
-    dailyNotePath,
-    currentDate
-  );
-  await focusDailyNote(dailyNotePath, isNew);
-}
-function getDailyNotePath(configuration: WorkspaceConfiguration, date: Date) {
-  const rootDirectory = workspace.workspaceFolders[0].uri.fsPath;
-  const dailyNoteDirectory: string =
-    configuration.get("openDailyNote.directory") ?? ".";
-  const dailyNoteFilename = getDailyNoteFileName(configuration, date);
-
-  return join(rootDirectory, dailyNoteDirectory, dailyNoteFilename);
-}
-
-function getDailyNoteFileName(
-  configuration: WorkspaceConfiguration,
-  date: Date
-): string {
-  const filenameFormat: string = configuration.get(
-    "openDailyNote.filenameFormat"
-  );
-  const fileExtension: string = configuration.get(
-    "openDailyNote.fileExtension"
-  );
-
-  return `${dateFormat(date, filenameFormat, false)}.${fileExtension}`;
-}
-
-async function createDailyNoteIfNotExists(
-  configuration: WorkspaceConfiguration,
-  dailyNotePath: string,
-  currentDate: Date
-) {
-  if (await pathExists(dailyNotePath)) {
-    return false;
-  }
-
-  await createDailyNoteDirectoryIfNotExists(dailyNotePath);
-
-  const titleFormat: string =
-    configuration.get("openDailyNote.titleFormat") ??
-    configuration.get("openDailyNote.filenameFormat");
-
-  await fs.promises.writeFile(
-    dailyNotePath,
-    `# ${dateFormat(currentDate, titleFormat, false)}${docConfig.eol}${
-      docConfig.eol
-    }`
-  );
-
-  return true;
-}
-
-async function createDailyNoteDirectoryIfNotExists(dailyNotePath: string) {
-  const dailyNoteDirectory = dirname(dailyNotePath);
-
-  if (!(await pathExists(dailyNoteDirectory))) {
-    await fs.promises.mkdir(dailyNoteDirectory, { recursive: true });
-  }
-}
-
-async function focusDailyNote(dailyNotePath: string, isNewNote: boolean) {
-  const document = await workspace.openTextDocument(Uri.file(dailyNotePath));
-  const editor = await window.showTextDocument(document);
-
-  // Move the cursor to end of the file
-  if (isNewNote) {
-    const { lineCount } = editor.document;
-    const { range } = editor.document.lineAt(lineCount - 1);
-    editor.selection = new Selection(range.end, range.end);
-  }
-}
-
-function pathExists(path: string) {
-  return fs.promises
-    .access(path, fs.constants.F_OK)
-    .then(() => true)
-    .catch(() => false);
-}
 
 export default feature;
